requirements.txt
praw
openai
python-dotenv
tqdm 
requests 
asyncpraw
aiohttp 